fix(api): ROUND に double precision を渡さないようにし、アナリティクスタブの500を解消 (#764)
PR #763 で chat_sessions.started_at の誤りを直した結果、同じクエリの
二段目のエラーが表面化した。本番ログで確定:

  DatabaseError: function round(double precision, integer) does not exist
  code: 42883
  at dist/src/api/admin/tenants/analyticsSummaryRoutes.js:66

COUNT(*) / GREATEST($2::float, 1) は double precision になるが、
PostgreSQL の2引数 round() は numeric 版しか存在しない。
numeric へキャストしてから丸めるよう修正する。

  ROUND(COUNT(*) / GREATEST($2::float, 1), 2)
  → ROUND((COUNT(*) / GREATEST($2::float, 1))::numeric, 2)

## なぜ1回で気づけなかったか

PostgreSQL は文のパース時に最初に見つけたエラーだけを返す。
created_at の列不存在が解決するまで round() の型エラーは表に出なかった。
同一クエリ内に独立した不具合が2つあった。

## 検証

PostgreSQL 17 で実測:
  - 修正前 → `function round(double precision, integer) does not exist` を再現
  - 修正後 → 0.07 が返る
  - 0件のとき → 0.00(壊れない)
  - days=0 のとき → GREATEST(...,1) が効き0除算にならない

## 他に同種が無いか確認

SQL内の2引数 ROUND は analytics/routes.ts に3箇所あるが、いずれも
`::numeric` を先にキャストしており正常。JS側の Math.round は無関係。
本件の1箇所のみが誤りだった。

## テスト

analyticsSummaryRoutes.test.ts に1件追加。ROUND の第1引数が ::numeric で
キャストされていること、float を直接渡す形に戻っていないことを固定する。
戻すと落ちることをミューテーションで確認済み。

## 補足(自分のミス)

当初この修正のコメントをテンプレートリテラル内にバッククォート付きで書き、
SQL文字列を途中で終端させてしまった(テストが ReferenceError で検出)。
テンプレートリテラル内のコメントでバッククォートを使わないこと。

+++ src/api/admin/tenants/analyticsSummaryRoutes.ts
@@ -75,7 +75,10 @@ export function registerAnalyticsSummaryRoutes(app: Express, db: Pool): void {
           db.query<{ total: string; avg_per_day: string }>(
             `SELECT
                COUNT(*)::text AS total,
-               ROUND(COUNT(*) / GREATEST($2::float, 1), 2)::text AS avg_per_day
+               -- COUNT(*)/float は double precision になるが、PostgreSQL の2引数 round() は
+               -- numeric 版しか無く "function round(double precision, integer) does not exist"
+               -- で落ちる。numeric へキャストしてから丸める。
+               ROUND((COUNT(*) / GREATEST($2::float, 1))::numeric, 2)::text AS avg_per_day
              FROM chat_sessions
              WHERE tenant_id = $1
                -- chat_sessions に created_at は存在しない。セッション開始時刻は started_at
